uscope/gstwidget.py
